Move compatibility and future behaviors into separate packages.
