+++ importlib_resources/_common.py
@@ -12,7 +12,7 @@ import itertools
 from typing import Union, Optional, cast
 from .abc import ResourceReader, Traversable
 
-from ._compat import wrap_spec
+from .future.adapters import wrap_spec
 
 Package = Union[types.ModuleType, str]
 Anchor = Package
